% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphanum_data.R
\docType{data}
\name{alphanum_data}
\alias{alphanum_data}
\title{Random alphanumeric data}
\format{A data frame with 5 observations of 3 numeric variables and 2 string variables:
\describe{
  \item{a}{random letters}
  \item{b}{random letters}
  \item{w}{random numbers with positive exponents}
  \item{y}{random numbers with negative exponents}
  \item{z}{random numbers with exponents = 0, 1, or 2}
  ...
}}
\usage{
alphanum_data
}
\description{
A dataset containing random character strings and numbers
}
\keyword{datasets}

