% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{density_data}
\alias{density_data}
\title{Air density data}
\format{A data frame with 5 observations of 3 variables:
\describe{
  \item{T_C}{air temperature, in C}
  \item{p_kPa}{air pressure, in kPa}
  \item{R}{gas constant, in J/(kg K)}
  ...
}}
\usage{
density_data
}
\description{
A dataset containing temperature and pressure measurements.
}
\keyword{datasets}

